refactor(utils): put `package_outputs` into a separate module
